Refactors and deduplications; improve diff output; update docs
