+++ sqlmesh/core/environment.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-from dataclasses import dataclass
 import json
 import re
 import typing as t
@@ -9,8 +8,10 @@ from pydantic import Field
 
 from sqlmesh.core import constants as c
 from sqlmesh.core.config import EnvironmentSuffixTarget
+from sqlmesh.core.engine_adapter.base import EngineAdapter
 from sqlmesh.core.macros import RuntimeStage
-from sqlmesh.core.snapshot import SnapshotId, SnapshotTableInfo
+from sqlmesh.core.renderer import render_statements
+from sqlmesh.core.snapshot import SnapshotId, SnapshotTableInfo, Snapshot
 from sqlmesh.utils import word_characters_only
 from sqlmesh.utils.date import TimeLike, now_timestamp
 from sqlmesh.utils.metaprogramming import Executable
@@ -213,21 +214,10 @@ class Environment(EnvironmentNamingInfo):
         return value if isinstance(value[0], dict) else [v.dict() for v in value]
 
 
-@dataclass
-class EnvironmentStatements:
+class EnvironmentStatements(PydanticModel):
     before_all: t.List[str]
     after_all: t.List[str]
     python_env: t.Dict[str, Executable]
-
-    @staticmethod
-    def parse_obj(statements: t.Dict[str, t.Any]) -> EnvironmentStatements:
-        return EnvironmentStatements(
-            before_all=statements.get(RuntimeStage.BEFORE_ALL.value, []),
-            after_all=statements.get(RuntimeStage.AFTER_ALL.value, []),
-            python_env={
-                key: Executable(**value) for key, value in statements.get("python_env", {}).items()
-            },
-        )
 
     def to_dict(self) -> t.Dict[str, t.Any]:
         return {
@@ -235,3 +225,34 @@ class EnvironmentStatements:
             RuntimeStage.AFTER_ALL.value: self.after_all,
             "python_env": {key: value.dict() for key, value in self.python_env.items()},
         }
+
+
+def execute_environment_statements(
+    adapter: EngineAdapter,
+    environment_statements: t.List[EnvironmentStatements],
+    runtime_stage: RuntimeStage,
+    environment_naming_info: EnvironmentNamingInfo,
+    default_catalog: t.Optional[str] = None,
+    snapshots: t.Optional[t.Dict[str, Snapshot]] = None,
+    start: t.Optional[TimeLike] = None,
+    end: t.Optional[TimeLike] = None,
+    execution_time: t.Optional[TimeLike] = None,
+) -> None:
+    if rendered_expressions := [
+        expr
+        for statements in environment_statements
+        for expr in render_statements(
+            statements=getattr(statements, runtime_stage.value),
+            dialect=adapter.dialect,
+            default_catalog=default_catalog,
+            python_env=statements.python_env,
+            snapshots=snapshots,
+            start=start,
+            end=end,
+            execution_time=execution_time,
+            environment_naming_info=environment_naming_info,
+        )
+    ]:
+        with adapter.transaction():
+            for expr in rendered_expressions:
+                adapter.execute(expr)
